App/requirement.txt
python-nmap
scapy
json
flask
flask-cors
psutil